% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfe_functions.R
\name{dfgm_martin}
\alias{dfgm_martin}
\title{DFE of FGM from Martin & Lenormand (2015).}
\source{
Martin, G., & Lenormand, T. (2015). The fitness effect of mutations
across environments: Fisher's geometrical model with multiple optima.
Evolution, 69(6), 1433-1447.
}
\usage{
dfgm_martin(s, n, lambda, so)
}
\arguments{
\item{s}{Vector of real numbers. Quantiles of the density. The density is 0
if \eqn{s \ge so}.}

\item{n}{Natural number. The dimensionality, i.e. the number of phenotypic
dimensions (traits) under selection.}

\item{lambda}{Positive real number. Mutational variance per trait among random
mutations scaled by the strength of selection. It scales the mean fitness effect
of mutations.}

\item{so}{Positive real number. Fitness distance between the wild-type (phenotype
without mutations) and the optimum.}
}
\value{
The density for each quantiles (selection coefficients) in a vector
format. The length of the vector is equal to the length of \code{s}.
}
\description{
Density of the distribution of fitness effects of random mutations from \href{https://doi.org/10.1111/evo.12671}{Martin
& Lenormand (2015)}. The density corresponds to equation (2) in the article:
\deqn{f_s(s, n, \lambda, s_o) = \frac{2}{\lambda}%
f_{\chi_{n}^2} (\frac{2 (s_o-s)}{\lambda}, \frac{2 s_o}{\lambda})}{%
fs(s, n, \lambda, so) = 2/\lambda d\chi[2,n](2 (so - s) / \lambda, 2 so / \lambda)}
The density is encoded using the functions from the package \code{\link[distr]{distr}}
}
\examples{
s <- seq(-2, 1, 0.1)
fs <- dfgm_martin(s, 2, 0.1, 1)

}
