% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfe.R
\name{dfgm_tenaillon}
\alias{dfgm_tenaillon}
\title{DFE of FGM from Tenaillon (2014).}
\source{
\itemize{
  \item Tenaillon, O. (2014). The utility of Fisher's geometric model in
  evolutionary genetics. Annual review of ecology, evolution, and systematics,
  45, 179-201.
  \item Martin, G., & Lenormand, T. (2015). The fitness effect of mutations
  across environments: Fisher's geometrical model with multiple optima.
  Evolution, 69(6), 1433-1447.
  \item Tenaillon, O., Silander, O. K., Uzan, J. P., & Chao, L. (2007).
  Quantifying organismal complexity using a population genetic approach.
  PloS one, 2(2), e217.
}
}
\usage{
dfgm_tenaillon(s, n, lambda, so, alpha, Q)
}
\arguments{
\item{s}{Vector of real numbers. Quantiles of the density. The density is 0
if \eqn{s \ge so}.}

\item{n}{Natural number. The dimensionality, i.e. the number of phenotypic
dimensions (traits) under selection.}

\item{lambda}{Positive real number. Mutational variance per trait among random
mutations scaled by the strength of selection. It scales the mean fitness effect
of mutations.}

\item{so}{Positive real number. Fitness distance between the wild-type (phenotype
without mutations) and the optimum.}

\item{alpha}{Positive real number. Scaling parameter in the formulation of
Tenaillon et al. (2007). (Identifiability problem with \code{lambda})}

\item{Q}{Positive real number. Parameter that modifies the concavity of the
fitness decline from the optimum.}
}
\value{
The density for each quantiles (selection coefficients) in a vector
format. The length of the vector is equal to the length of \code{s}.
}
\description{
Density of the distribution of fitness effects of random mutations from
\href{https://doi.org/10.1146/annurev-ecolsys-120213-091846}{Tenaillon (2014)}
adapted to show the same parameters as Martin & Lenormand (2015) plus the
parameters \code{Q} and \code{alpha}.
The density corresponds to the first equation in the paragraph *"Distribution of
mutation effects"* in the article.
The density is encoded using the functions from the package \code{\link[gsl]{gsl}}
}
\examples{
s <- sample(seq(-2, 1.5, 0.1))
fs <- dfgm_tenaillon(s, 2, 0.1, 1, 1/2, 2)

}
