% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_function.R
\name{rfgm}
\alias{rfgm}
\title{Random generation of selection coefficient under FGM}
\source{
\itemize{
  \item Tenaillon, O. (2014). The utility of Fisher's geometric model in
  evolutionary genetics. Annual review of ecology, evolution, and systematics,
  45, 179-201.
  \item Martin, G., & Lenormand, T. (2015). The fitness effect of mutations
  across environments: Fisher's geometrical model with multiple optima.
  Evolution, 69(6), 1433-1447.
 }
}
\usage{
rfgm(nsample, n, lambda, so, alpha, Q)
}
\arguments{
\item{nsample}{Real number. Number of mutations to draw.}

\item{n}{Natural number. The dimensionality, i.e. the number of phenotypic
dimensions (traits) under selection.}

\item{lambda}{Positive real number. Mutational variance per trait among random
mutations scaled by the strength of selection. It scales the mean fitness effect
of mutations.}

\item{so}{Positive real number. Fitness distance between the wild-type (phenotype
without mutations) and the optimum.}

\item{alpha}{Positive real number. Scaling parameter in the formulation of
Tenaillon et al. (2007). (Identifiability problem with \code{lambda})}

\item{Q}{Positive real number. Parameter that modifies the concavity of the
fitness decline from the optimum.}
}
\value{
Vector of selection coefficients of length \code{nsample}.
}
\description{
Generates selection coefficients of random mutations. First, the phenotype are
drawn from a multinormal distribution with means : 0, variances : lambda, and
covariances : 0. Second the coefficient of selections of these phenotypes are
computed using Fisher's Geometric Model (see Tenaillon (2014), Martin & Lenormand
(2015)).
}
\examples{
rfgm(10, 2, 0.1, 1, 1/2, 2)

}
